[Strainer] Separate factory from template
